Move Legacy impl into /domain
